Fix compiler warning: implicit conversion loses integer precision
src/utils_files.c:268:28: warning: implicit conversion loses integer precision: 'unsigned long' to 'gint' (aka 'int') [-Wshorten-64-to-32]
                long_str = strlen (string);
                         ~ ^~~~~~~~~~~~~~~
src/utils_files.c:916:16: warning: implicit conversion loses integer precision: 'unsigned long' to 'gint' (aka 'int') [-Wshorten-64-to-32]
        dirstr_len  = strlen (dirstr);
                    ~ ^~~~~~~~~~~~~~~
src/utils_files.c:919:16: warning: implicit conversion loses integer precision: 'unsigned long' to 'gint' (aka 'int') [-Wshorten-64-to-32]
        sepstr_len  = strlen (sepstr);
                    ~ ^~~~~~~~~~~~~~~

+++ src/utils_files.c
@@ -240,7 +240,7 @@ static GSList *utils_files_check_UTF8_validity (const gchar *contents,
     GSList *list = NULL;
     struct struc_check_encoding *result;
     gchar *string = NULL;
-    gint long_str = 0;
+    gsize long_str = 0;
     gsize size = 0;
     gsize bytes_written = 0;
     gint i = 0;
@@ -904,9 +904,9 @@ gchar *utils_files_selection_get_last_directory (GtkFileChooser *filesel,
 											     gboolean ended)
 {
     gchar *dirstr;
-    gint dirstr_len = 0;
+    gsize dirstr_len = 0;
     gchar *sepstr;
-    gint sepstr_len = 0;
+    gsize sepstr_len = 0;
     gboolean is_endedstr = FALSE;
     gchar* tmpstr;
